refactor: rename module line by cmdline
